ref: Move browser code from __init__.py into browser.py
